[device-bootloaders] Multiple bug fixes

+++ device-bootloaders/src/ServiceProcessorBL2/include/minion_configuration.h
@@ -320,12 +320,12 @@ int8_t enable_sram_and_icache_interrupts(void);
 */
 int8_t disable_sram_and_icache_interrupts(void);
 
-/*! \fn int8_t Minion_VPU_RF_Init(uint8_t shire_id) 
+/*! \fn int Minion_VPU_RF_Init(uint8_t shire_id) 
  *  \brief This function initialized the VPU RF for all Minion.
  *         listed on the shire_mask
  *  \param shire_id Minion Shire ID
  *  \return Status indicating success or negative error
 */
-int8_t Minion_VPU_RF_Init(uint8_t shireid);
+int Minion_VPU_RF_Init(uint8_t shireid);
 
 #endif
